Added basic Cython header file for rmg.data module.
To get the rmg.data module to compile, I had to move a lambda function to be a
new private member function in the Database class. I also had to fix a bit of
outdated code (which would not have run anyway).

Some typing of parameters has been added. The modifications to the rmg.io
module are to ensure that these methods get passed str objects and not unicode
objects.