Added base HTML template and default stylesheet.
The current template provides two columns: a left sidebar and a right
main area with a header, navigation bar, contents area, and footer. The
styling is inspired by Wikipedia, and colorized with MIT colors,
particularly maroon. The color styling is collected in the same area of
the stylesheet for easy changing.