First version of a form for adding thermodynamics entries to the database.
Added forms module to the database app for definition of the various
forms in the app. Note that we will generally not base these on models,
as we will use our own classes for these (the same ones needed for RMG
Py).

Added url handlers for adding thermo, kinetics, and states entries. Only
the thermo handler leads to a view with a valid rendering at the moment.