Much improved form validation for adding thermo entries to database.
Clicking "Submit" on the form now causes a large amount of validation to
proceed. If the form is deemed valid, an object containing the cleaned data
is created. This object is ready for adding to the database.

Key takeaway: Processing of the POST command in the appropriate view should
use only the *cleaned* data, i.e. the data should already be in a form
suitable for adding to the database at this point. This means that type and
unit conversions should be done in other places.