Added groundwork for user authentication system.
We will want to be able to authenticate users, both so that not just anyone
can spam our web interface, and that we can associate submitted changes with
the user that submitted them. Viewing of the database will be open to all
users, but adding and manipulating the database will require a user account.

There is currently no facility for getting a user account. This will come
later.