Added first version of new ThermoDepository and ThermoDatabase classes.
Previously, ThermoDatabase was used as a base class for the group and primary
thermo databases. Now I wish to repurpose it as a container class for all
loaded thermo information: depository, library, and groups. I have renamed the
old ThermoDatabase to ThermoDatabaseBase for now to avoid a name conflict.

The ThermoDepository class represents a depository of thermodynamic data. The
depository associates species with thermodynamic parameters. So far I've
written a rudimentary function for saving records to the depository.