Added first version of thermo evaluation form for thermo entry viewing.
This form uses Javascript (via jQuery) to dynamically update the
ThermoEvalForm in response to the user modifying the temperature field or any
of the units fields. This unfortunately requires implementation of thermo
parameter evaluation functions in Javascript, as there isn't a good way to
connect this to Python. The Javascript file chem.js was created to hold
functions such as these. Again only Wilhoit models have been implemented so
far, with others to follow.