Thermo entry viewer now works for GA, Wilhoit, and NASA thermo models.
The template can handle any number of Cp(T) points supplied to the
ThermoGAModel and any number of NASA polynomials. This is also true of
the thermo evaluator tool. The parameters for each model are now
displayed using a two-column table.

The Javascript code for the thermo evaluator has been moved from the
viewThermoEntry template to the thermoEvalForm template. This way we can
potentially use this template in other places as well.