Added KineticsFamily.getRateCoefficientUnits() method.
This provides a way to determine the typical units of a rate coefficient
for that family: s^-1 if unimolecular or m^3/(mol*s) if bimolecular.
The current implementation uses hardcoded lists that are available in
rmgpy.data.kinetics.common; perhaps a future implementation will find a
better way, but at least you can now call a method and not care about the
implementation details most of the time.