Fixed minor bug in ThermoData.getEntropy() for monatomics.
The bug only occurred if the heat capacity was constant (i.e. for
monatomics). It would cause a NaN to be returned if you tried to ask for
an entropy value outside of the range covered by Tdata, as in that case
slope would evaluate to zero if CpInf == Cpdata[-1].