Added compressed sparse row format for Graph class.
The compressed sparse row (CSR) format is designed to use much less memory
than the dict of dicts (DoD) format. CSR may also be slightly faster for
traversing the graph if you use integer indices instead of Vertex and Edge
objects, though that is not completely clear. DoD is certainly the better
format for editing the graph, so that functionality remains. The Graph class
should convert between the two representations silently as needed, though we
will need to be sure to avoid doing this too much since it costs some CPU
time.