Updated implementation of VF2 algorithm to use new CSR Graph format.
The key change here is that graph isomorphism information is no longer
cached on the Vertex or Edge objects. Connectivity information is now stored
on the Graph class, while mapping and terminal information is stored on the
VF2 class (since they only have meaning in the VF2 algorithm).

This commit (and the previous one) introduce the use of Cython's typed
memoryviews feature, which gives a cleaner, faster syntax for things like
types arrays, but requires Cython 0.15 or later.