Updated Molecule and Graph classes to work with new Graph class.
The Atom and GroupAtom classes were updated to remove the bonds attribute,
since we no longer store the edges on the Vertex object. Similarly, the Bond
and GroupBond classes were updated to remove the atom1 and atom2 attributes,
since we no longer store the vertices on the Edge object. Most of the
remainder of this commit is simply updating the references to those
attributes accordingly.