Fix AtomTypeTest.testSetActions() to not overwrite the R atom type.
This tests the modifying of the various attributes of the AtomType class.
Previously it used the R atom type for this, which caused its properties to
change to nonsensical values, which broke subsequent unit tests. Now we use
a blank AtomType object that we are free to modify without consequence.