Restoring support for storing hydrogens implicitly in Molecule objects.
We used this for a while before, but weren't sure that we were seeing enough
benefit to justify the additional code complexity. Now that things are
working a little better, we spend a lot of time making copies of things. I'm
thinking this can be reduced if we no longer store most of the hydrogen
atoms. This may also help get the memory requirements closer to that of RMG-
Java.

Unlike before, this implementation makes use of implicit hydrogens the
default state, which simplifies the code a bit. The implementation is not
optimized for CPU time (yet), as I'm intending to test the memory use first.